Move services to modules
