+++ src/components/MainPage/Importers/Importer.vue
@@ -45,7 +45,7 @@ import { MessageBox } from 'element-ui';
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { remote } from 'electron';
 import { decryptProperty } from '@/modules/encryption/credentials';
-import { scrape } from '@/service/scrapers.service';
+import { scrape } from '@/modules/scrapers';
 
 export default {
   props: {
